requirements.txt
numpy~=1.23.5
pandas==1.4.3
pandas_stubs==1.2.0.56
Pillow~=9.3.0
requests~=2.28.1
streamlit ==1.11.1
tensorflow==2.9.1

pyttsx3~=2.90
wikipedia~=1.4.0
openai~=0.26.5
a~=1.0
opencv-python~=4.6.0.66
pytesseract ~=0.3.10
gTTS ~=2.3.0
playsound ~=1.3.0